extract delegates, IStore and add simple explanations
